Star Wars API - Starships coding challenge
The solution meets the requirements of the coding challenge by means of
several asynchronous calls.  Once each response is received, the number
off re-supplies required for each star-ship to travel the desired
distance is calculated.  There are also a number of unit test include
which assert various inputs, valid or otherwise.  In addition, I have
included a test to verify that the URL supplied returns a successful
response.